server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==26.1.0
billiard==4.2.4
boto3==1.36.0
botocore==1.36.26
celery==5.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
flower==2.0.1
frozenlist==1.8.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
humanize==4.15.0
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-experimental==0.4.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.29
lightgbm==4.5.0
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==6.1.0
numpy==2.4.4
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pinecone==8.1.2
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
prometheus_client==0.21.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pypdf==5.4.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
pytz==2026.1.post1
PyYAML==6.0.3
redis==5.2.1
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
s3transfer==0.11.3
scikit-learn==1.6.1
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tornado==6.5.5
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
slowapi==0.1.9
limits==3.13.0
